+++ reqwest-retry/src/middleware.rs
@@ -6,7 +6,7 @@ use chrono::Utc;
 use reqwest::{Request, Response};
 use reqwest_middleware::{Error, Middleware, Next, Result};
 use retry_policies::RetryPolicy;
-use truelayer_extensions::Extensions;
+use task_local_extensions::Extensions;
 
 /// We limit the number of retries to a maximum of `10` to avoid stack-overflow issues due to the recursion.
 static MAXIMUM_NUMBER_OF_RETRIES: u32 = 10;